+++ src/http_util/body.rs
@@ -22,7 +22,6 @@ use std::pin::Pin;
 use std::task::Context;
 use std::task::Poll;
 
-use anyhow::anyhow;
 use bytes::Bytes;
 use futures::channel::mpsc;
 use futures::channel::mpsc::Sender;
@@ -30,17 +29,15 @@ use futures::ready;
 use futures::AsyncWrite;
 use futures::SinkExt;
 use futures::StreamExt;
-use http::response::Parts;
-use http::Response;
 use http::StatusCode;
-use hyper::body::HttpBody;
 use hyper::client::ResponseFuture;
 use hyper::Body;
 use log::debug;
 use pin_project::pin_project;
 
 use crate::error::other;
-use crate::error::ObjectError;
+use crate::http_util::error::ParseErrorResponse;
+use crate::http_util::parse_error_response;
 use crate::ops::OpWrite;
 
 /// Create a HTTP channel.
@@ -68,7 +65,7 @@ enum State {
 }
 
 impl HttpBodyWriter {
-    /// Create a HTTP body writer.
+    /// Create an HTTP body writer.
     ///
     /// # Params
     ///
@@ -156,87 +153,13 @@ impl AsyncWrite for HttpBodyWriter {
     }
 }
 
-/// parse_error_response will try to read and parse error response.
-pub fn parse_error_response(
-    op: &'static str,
-    path: &str,
-    parser: fn(StatusCode) -> ErrorKind,
-    resp: Response<Body>,
-) -> ParseErrorResponse {
-    let (parts, body) = resp.into_parts();
-
-    ParseErrorResponse {
-        op,
-        path: path.to_string(),
-        parser,
-        parts,
-        body,
-        buf: Vec::with_capacity(1024),
-    }
-}
-
-pub struct ParseErrorResponse {
-    op: &'static str,
-    path: String,
-    parser: fn(StatusCode) -> ErrorKind,
-    parts: Parts,
-    body: Body,
-
-    buf: Vec<u8>,
-}
-
-impl Future for ParseErrorResponse {
-    type Output = Error;
-
-    fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        match ready!(Pin::new(&mut self.body).poll_data(cx)) {
-            None => Poll::Ready(Error::new(
-                (self.parser)(self.parts.status),
-                ObjectError::new(
-                    self.op,
-                    &self.path,
-                    anyhow!(
-                        "status code: {:?}, headers: {:?}, body: {:?}",
-                        self.parts.status,
-                        self.parts.headers,
-                        String::from_utf8_lossy(&self.buf)
-                    ),
-                ),
-            )),
-            Some(Ok(data)) => {
-                // Only read 4KiB from the response to avoid broken services.
-                if self.buf.len() < 4 * 1024 {
-                    self.buf.extend_from_slice(&data);
-                }
-
-                // Make sure the whole body consumed, even we don't need them.
-                self.poll(cx)
-            }
-            Some(Err(e)) => Poll::Ready(Error::new(
-                (self.parser)(self.parts.status),
-                ObjectError::new(
-                    self.op,
-                    &self.path,
-                    anyhow!(
-                        "status code: {:?}, headers: {:?}, read body: {:?}, remaining {:?}",
-                        self.parts.status,
-                        self.parts.headers,
-                        String::from_utf8_lossy(&self.buf),
-                        e
-                    ),
-                ),
-            )),
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use futures::SinkExt;
     use serde::Deserialize;
 
     use super::*;
-    use crate::io_util::HttpClient;
+    use crate::http_util::HttpClient;
 
     #[derive(Deserialize, Default)]
     #[serde(default)]
