+++ src/http_util/mod.rs
@@ -1,0 +1,32 @@
+// Copyright 2022 Datafuse Labs.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+mod body;
+pub use body::new_http_channel;
+pub use body::HttpBodyWriter;
+
+mod client;
+pub use client::HttpClient;
+
+mod header;
+pub use header::parse_content_length;
+pub use header::parse_content_md5;
+pub use header::parse_etag;
+pub use header::parse_last_modified;
+
+mod uri;
+pub use uri::percent_encode_path;
+
+mod error;
+pub use error::parse_error_kind;
+pub use error::parse_error_response;
