+++ src/http_util/uri.rs
@@ -12,83 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::io::ErrorKind;
-use std::ops::Deref;
-
-#[cfg(feature = "rustls")]
-use hyper_rustls::{HttpsConnector, HttpsConnectorBuilder};
-#[cfg(not(feature = "rustls"))]
-use hyper_tls::HttpsConnector;
-
-use percent_encoding::utf8_percent_encode;
-use percent_encoding::AsciiSet;
-use percent_encoding::NON_ALPHANUMERIC;
-
-/// HttpClient that used across opendal.
-///
-/// NOTE: we could change or support more underlying http backend.
-#[derive(Debug, Clone)]
-pub struct HttpClient(hyper::Client<HttpsConnector<hyper::client::HttpConnector>, hyper::Body>);
-
-#[cfg(not(feature = "rustls"))]
-#[inline]
-pub(crate) fn https_connector() -> HttpsConnector<hyper::client::HttpConnector> {
-    HttpsConnector::new()
-}
-
-#[cfg(feature = "rustls")]
-#[inline]
-pub(crate) fn https_connector() -> HttpsConnector<hyper::client::HttpConnector> {
-    HttpsConnectorBuilder::new()
-        .with_native_roots()
-        .https_or_http()
-        .enable_http1()
-        .build()
-}
-
-impl HttpClient {
-    /// Create a new http client.
-    pub fn new() -> Self {
-        HttpClient(
-            hyper::Client::builder()
-                // Disable connection pool to address weird async runtime hang.
-                //
-                // ref: https://github.com/datafuselabs/opendal/issues/473
-                .pool_max_idle_per_host(0)
-                .build(https_connector()),
-        )
-    }
-}
-
-impl Default for HttpClient {
-    fn default() -> Self {
-        HttpClient::new()
-    }
-}
-
-/// Forward all function to http backend.
-impl Deref for HttpClient {
-    type Target = hyper::Client<HttpsConnector<hyper::client::HttpConnector>, hyper::Body>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-/// Parse hyper error into `ErrorKind`.
-///
-/// It's safe to retry the following errors:
-///
-/// - is_canceled(): Request that was canceled.
-/// - is_connect(): an error from Connect.
-/// - is_timeout(): the error was caused by a timeout.
-pub fn parse_error_kind(err: &hyper::Error) -> ErrorKind {
-    if err.is_canceled() || err.is_connect() || err.is_timeout() {
-        ErrorKind::Interrupted
-    } else {
-        ErrorKind::Other
-    }
-}
+use percent_encoding::{utf8_percent_encode, AsciiSet, NON_ALPHANUMERIC};
 
 /// PATH_ENCODE_SET is the encode set for http url path.
 ///
