+++ src/lib.rs
@@ -128,4 +128,5 @@ pub mod services;
 //
 // Please don't export any type from this module.
 mod error;
+mod http_util;
 mod path;
