+++ src/services/http/backend.rs
@@ -40,16 +40,16 @@ use radix_trie::TrieCommon;
 use crate::error::other;
 use crate::error::BackendError;
 use crate::error::ObjectError;
-use crate::io_util::new_http_channel;
-use crate::io_util::parse_content_length;
-use crate::io_util::parse_content_md5;
-use crate::io_util::parse_error_kind as parse_http_error_kind;
-use crate::io_util::parse_error_response;
-use crate::io_util::parse_etag;
-use crate::io_util::parse_last_modified;
-use crate::io_util::percent_encode_path;
-use crate::io_util::HttpBodyWriter;
-use crate::io_util::HttpClient;
+use crate::http_util::new_http_channel;
+use crate::http_util::parse_content_length;
+use crate::http_util::parse_content_md5;
+use crate::http_util::parse_error_kind as parse_http_error_kind;
+use crate::http_util::parse_error_response;
+use crate::http_util::parse_etag;
+use crate::http_util::parse_last_modified;
+use crate::http_util::percent_encode_path;
+use crate::http_util::HttpBodyWriter;
+use crate::http_util::HttpClient;
 use crate::ops::BytesRange;
 use crate::ops::OpCreate;
 use crate::ops::OpDelete;
@@ -157,7 +157,7 @@ impl Builder {
     }
 
     /// Build a HTTP backend.
-    pub async fn finish(&mut self) -> Result<Arc<dyn Accessor>> {
+    pub fn build(&mut self) -> Result<Backend> {
         info!("backend build started: {:?}", &self);
 
         let endpoint = match &self.endpoint {
@@ -194,12 +194,18 @@ impl Builder {
         let client = HttpClient::new();
 
         info!("backend build finished: {:?}", &self);
-        Ok(Arc::new(Backend {
+        Ok(Backend {
             endpoint: endpoint.to_string(),
             root,
             client,
             index: Arc::new(Mutex::new(mem::take(&mut self.index))),
-        }))
+        })
+    }
+
+    /// Build a HTTP backend.
+    #[deprecated = "Use Builder::build() instead"]
+    pub async fn finish(&mut self) -> Result<Arc<dyn Accessor>> {
+        Ok(Arc::new(self.build()?))
     }
 }
 
@@ -231,13 +237,12 @@ impl Debug for Backend {
 
 impl Backend {
     /// Create a new builder for s3.
+    #[deprecated = "Use Builder::default() instead"]
     pub fn build() -> Builder {
         Builder::default()
     }
 
-    pub(crate) async fn from_iter(
-        it: impl Iterator<Item = (String, String)>,
-    ) -> Result<Arc<dyn Accessor>> {
+    pub(crate) fn from_iter(it: impl Iterator<Item = (String, String)>) -> Result<Self> {
         let mut builder = Builder::default();
 
         for (k, v) in it {
@@ -249,7 +254,7 @@ impl Backend {
             };
         }
 
-        builder.finish().await
+        builder.build()
     }
 
     pub(crate) fn get_abs_path(&self, path: &str) -> String {
