+++ src/services/s3/dir_stream.rs
@@ -32,7 +32,7 @@ use serde::Deserialize;
 use super::Backend;
 use crate::error::other;
 use crate::error::ObjectError;
-use crate::io_util::parse_error_response;
+use crate::http_util::parse_error_response;
 use crate::services::s3::backend::parse_error_kind;
 use crate::DirEntry;
 use crate::ObjectMode;
